separate streams and video
